+++ eras/babbage/impl/src/Cardano/Ledger/Babbage.hs
@@ -16,6 +16,7 @@ module Cardano.Ledger.Babbage (
   BabbageTxBody,
   AlonzoScript,
   AlonzoTxAuxData,
+  getDatumBabbage,
 )
 where
 
@@ -23,6 +24,8 @@ import Cardano.Ledger.Alonzo (reapplyAlonzoTx)
 import Cardano.Ledger.Alonzo.Genesis (AlonzoGenesis (..))
 import Cardano.Ledger.Alonzo.Language (Language (..))
 import Cardano.Ledger.Alonzo.Scripts (AlonzoScript (..))
+import Cardano.Ledger.Alonzo.Scripts.Data (Data)
+import Cardano.Ledger.Alonzo.Tx (AlonzoEraTx, ScriptPurpose)
 import Cardano.Ledger.Alonzo.TxAuxData (AlonzoTxAuxData (..))
 import Cardano.Ledger.Alonzo.TxInfo (EraPlutusContext, ExtendedUTxO (..))
 import Cardano.Ledger.Babbage.Core
@@ -33,7 +36,6 @@ import Cardano.Ledger.Babbage.Transition ()
 import Cardano.Ledger.Babbage.Translation ()
 import Cardano.Ledger.Babbage.Tx (
   babbageTxScripts,
-  getDatumBabbage,
  )
 import Cardano.Ledger.Babbage.TxAuxData ()
 import Cardano.Ledger.Babbage.TxBody (
@@ -41,7 +43,7 @@ import Cardano.Ledger.Babbage.TxBody (
   BabbageTxOut,
  )
 import Cardano.Ledger.Babbage.TxInfo (babbageTxInfo)
-import Cardano.Ledger.Babbage.UTxO ()
+import Cardano.Ledger.Babbage.UTxO (getBabbageSpendingDatum)
 import Cardano.Ledger.Binary (sizedValue)
 import Cardano.Ledger.Crypto (Crypto, StandardCrypto)
 import Cardano.Ledger.Keys (DSignable, Hash)
@@ -76,4 +78,16 @@ instance (Crypto c, EraPlutusContext 'PlutusV2 (BabbageEra c)) => ExtendedUTxO (
       newOuts = map sizedValue $ toList $ txBody ^. allSizedOutputsTxBodyF
       referencedOuts = Map.elems $ Map.restrictKeys utxo (txBody ^. referenceInputsTxBodyL)
       outs = newOuts <> referencedOuts
-  getDatum = getDatumBabbage
+
+-- | Extract binary data either directly from the `Tx` as an "inline datum"
+-- or look it up in the witnesses by the hash.
+getDatumBabbage ::
+  ( AlonzoEraTx era
+  , BabbageEraTxOut era
+  ) =>
+  Tx era ->
+  UTxO era ->
+  ScriptPurpose era ->
+  Maybe (Data era)
+getDatumBabbage tx utxo = getBabbageSpendingDatum utxo tx
+{-# DEPRECATED getDatumBabbage "In favor of `getBabbageSpendingDatum`" #-}