+++ eras/alonzo/impl/CHANGELOG.md
@@ -1,8 +1,10 @@
 # Version history for `cardano-ledger-alonzo`
 
-## 1.4.2.1
+## 1.5.0.0
 
-*
+* Add `AlonzoEraUTxO` and use in the rules.
+* Deprecate `getDatumAlonzo` in favor of `getAlonzoSpendingDatum`
+* Deprecate `getSpendingTxIn` in favor of `getAlonzoSpendingTxIn`
 
 ## 1.4.2.0